+++ src/elements/core/interfaces.ts
@@ -1,3 +1,4 @@
 export * from './interfaces/overlay-close-details.js';
+export * from './interfaces/paginator-page.js';
 export * from './interfaces/types.js';
 export * from './interfaces/validation-change.js';
