+++ src/elements/core/interfaces/paginator-page.ts
@@ -1,0 +1,6 @@
+export type SbbPaginatorPageEventDetails = {
+  length: number;
+  pageSize: number;
+  pageIndex: number;
+  previousPageIndex: number;
+};
